Add a verbosity config option, and a joined-channel message.
This is aimed at resolving #1452, but won't be complete without a
puppet change. The verbosity setting removes most of the chatter
that would otherwise be output, so add a ready-to-plub message when
we know we've joined the channel of interest. This makes it possible
to see that the bot is working from the command line.

Change-Id: I68e42d661884c38b34f30dcaf333bc64744dad23