Emit a warning if PIPE_PATH isn't a FIFO.
Replacing it with a file is a fairly typical workaround when pipebot
goes down for some reason. If we forget to restore the file before
bringing pipebot back up we get a cryptic error in the logs. This
ought to be much clearer.

Patch set 2: Also include the actual path checked..

Change-Id: Ie9a7e66ee0ad37804b156b794236ca2b3e494b13