Log all the things
Since SRBot is likely to go away at some point, and really isn't
doing very much for us anyway, we're going to need another thing
which logs the IRC channel. This commit provides a _very_ basic
logging mechanism to replace that.

If SRBot does go away, we might want to look at replacing it with
something which can do the other things we use it for, but logging
is the most critical of those, so just resolve that one first.

Change-Id: I951074da19678a67997f21416c17279aee8fb6d3